refactor rbrt to use f32 only to make simd easier

+++ rbrt_lib/src/aabbox.rs
@@ -8,11 +8,11 @@ pub struct BoundingBox {
 }
 
 /// just for better readabilty
-pub fn max(a: f64, b: f64) -> f64 {
+pub fn max(a: f32, b: f32) -> f32 {
     return a.max(b);
 }
 
-pub fn min(a: f64, b: f64) -> f64 {
+pub fn min(a: f32, b: f32) -> f32 {
     return a.min(b);
 }
 
@@ -58,11 +58,10 @@ impl BoundingBox {
     }
 }
 
-
 /// computes the axis aligned bounding box extents of triangles
 pub fn compute_min_max_3d(triangle_mesh: &Vec<[Vec3; 3]>) -> (Vec3, Vec3) {
-    let mut lower_bound_tmp = Vec3::new(std::f64::MAX, std::f64::MAX, std::f64::MAX);
-    let mut upper_bound_tmp = Vec3::new(-std::f64::MAX, -std::f64::MAX, -std::f64::MAX);
+    let mut lower_bound_tmp = Vec3::new(std::f32::MAX, std::f32::MAX, std::f32::MAX);
+    let mut upper_bound_tmp = Vec3::new(-std::f32::MAX, -std::f32::MAX, -std::f32::MAX);
     for tri in triangle_mesh {
         for idx in 0..3 {
             let vertex = tri[idx];
